pyproject.toml
[project]
name = "youtube_summarizer"
version = "0.1.0"
description = ""
authors = [{ name = "Git-Monke", email = "javelasq2006@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "pytube2 (>=15.0.16,<16.0.0)",
    "yt-dlp (>=2025.7.21,<2026.0.0)",
    "faster-whisper (>=1.1.1,<2.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sse-starlette (>=3.0.2,<4.0.0)",
    "tinydb (>=4.8.2,<5.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "ollama (>=0.5.1,<0.6.0)",
]

[tool.poetry]
packages = [{ include = "youtube_summarizer", from = "src" }]

[tool.poetry.scripts]
youtube-summarizer = "youtube_summarizer.__main__:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
